+++ Tests/EditorMode/Service/Save/SaveCapturerV2Tests.cs
@@ -1,8 +1,8 @@
 using System;
 using System.Collections.Generic;
-using System.Text.RegularExpressions;
 using Moirai.Atropos;
 using Moirai.Atropos.Save;
+using Moirai.Atropos.Tests.EditorMode;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -96,12 +96,7 @@ namespace Service.Save
         /// </summary>
         private static void ExpectWarningLogForUtf()
         {
-#if UNITY_LOGGING_INSTALLED
-            if (LogUtility.Handler is not UnityLoggingHandler)
-#endif
-            {
-                LogAssert.Expect(LogType.Warning, new Regex(".*"));
-            }
+            UtfLogExpect.Warning();
         }
 
         /// <summary>