+++ Tests/EditorMode/Service/Save/SaveKeyProviderTests.cs
@@ -1,10 +1,10 @@
 using System;
 using System.Collections.Generic;
 using System.IO;
-using System.Text.RegularExpressions;
 using System.Threading;
 using Moirai.Atropos;
 using Moirai.Atropos.Save;
+using Moirai.Atropos.Tests.EditorMode;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -83,12 +83,7 @@ namespace Service.Save
         /// </summary>
         private static void ExpectErrorLogForUtf()
         {
-#if UNITY_LOGGING_INSTALLED
-            if (LogUtility.Handler is not UnityLoggingHandler)
-#endif
-            {
-                LogAssert.Expect(LogType.Error, new Regex(".*"));
-            }
+            UtfLogExpect.Error();
         }
 
         private static AESEncryptedSaveHandler CreateHandler(string passphrase)