+++ Tests/EditorMode/Service/Save/SaveMigrationAndBackendTests.cs
@@ -7,12 +7,12 @@ using MessagePack;
 #endif
 using Moirai.Atropos;
 using Moirai.Atropos.Save;
+using Moirai.Atropos.Tests.EditorMode;
 using NUnit.Framework;
 #if PROTOBUF_INSTALLED
 using ProtoBuf;
 #endif
 using System.IO;
-using System.Text.RegularExpressions;
 using System.Threading;
 using UnityEngine;
 using UnityEngine.TestTools;
@@ -148,12 +148,7 @@ namespace Service.Save
         /// </summary>
         private static void ExpectErrorLogForUtf()
         {
-#if UNITY_LOGGING_INSTALLED
-            if (LogUtility.Handler is not UnityLoggingHandler)
-#endif
-            {
-                LogAssert.Expect(LogType.Error, new Regex(".*"));
-            }
+            UtfLogExpect.Error();
         }
 
         private SaveServiceHandler.SavePaths Paths(string fileName)